feat: #2 update dependencies
- Move user auth to dependencies
